Seperate a code into Files #12
