+++ lib/widgets/ScheduleInputModal.dart
@@ -3,6 +3,35 @@ import 'package:flutter_datetime_picker/flutter_datetime_picker.dart';
 import 'package:intl/intl.dart';
 import 'package:flutter_datetime_picker/src/date_format.dart';
 
+class SchduleInputModal extends StatefulWidget{
+  @override
+  _SchduleInputModalState createState() => _SchduleInputModalState();
+}
+
+class _SchduleInputModalState extends State<SchduleInputModal> {
+  @override
+  Widget build(BuildContext context) {
+    return  AlertDialog(
+      //title: Container(child: new Text('각 일정 탭 클릭')),
+      content:Container(
+          height: MediaQuery.of(context).size.height*0.5,
+          width: MediaQuery.of(context).size.width*0.8,
+          child: ScheduleInputForm()
+      ),
+      shape: RoundedRectangleBorder( // 모서리 둥글게
+          borderRadius: BorderRadius.all(Radius.circular(10.0))),
+      actions: <Widget>[
+        new TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: new Text('추가하기'),
+            style: TextButton.styleFrom(primary: Colors.black)
+        )],
+    );
+  }
+}
+
 class ScheduleInputForm extends StatefulWidget {
   @override
   _ScheduleInputFormState createState() => _ScheduleInputFormState();
